refactor: extract WeeklyPracticeRow to Components/
Preparatory move so JournalView can render the 7-day practice dots.
No behavior change.
